Added support for /3/person/latest
Created GetPersonItem method in TMDbClientPeople to support the
person/latest api get request.

Created an enumeration describing the PersonItemType (for now, we only
have 'Latest') - this resembles MovieListType / PersonListType.

Note that this is a different method to GetPersonList because /latest
only returns a single person, rather than a list of people (thus it is
not wrapped in a SearchContainer).

Added TestPersonItem to test this new method.