Add missing Person Translations
And I found another endpoint/method missing that I needed,
this time for the people. I did a PR for mapping the person bio.
to overview previously, but forgot to check if the methods/endpoints
to get the translation was mapped… which they were not… so here I
am again, with another PR. 😅